@@ -4,23 +4,22 @@
   "private": true,
   "dependencies": {
     "@types/node": "^16.11.26",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "i18next": "^21.6.14",
     "i18next-http-backend": "^1.4.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-i18next": "^11.16.2",
-    "react-icons": "^4.3.1",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.6.2"
+    "react-icons": "^4.3.1"
   },
   "scripts": {
-    "devstart": "react-scripts start",
-    "build": "react-scripts build",
+    "devstart": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
     "prettier:write": "prettier --write src/ public/ --ignore-path .prettierignore",
     "prettier:check": "prettier --check src/ public/ --ignore-path .prettierignore",
     "prettier:ci": "npx prettier --check src/ public/ --ignore-path .prettierignore",
@@ -45,6 +44,11 @@
   },
   "devDependencies": {
     "husky": "^7.0.0",
-    "prettier": "^2.6.0"
+    "prettier": "^2.6.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@vitejs/plugin-react": "^1.3.0",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.9"
   }
 }
