@@ -1,5 +1,5 @@
 import React, { StrictMode, Suspense } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import App from "./components/App";
 import "bootstrap/dist/css/bootstrap.min.css";
@@ -11,6 +11,8 @@ const fallbackUi = React.createElement("p", {}, "Loading...");
 const app = React.createElement(App, { key: "#" }, []);
 const strictApp = React.createElement(StrictMode, { key: "strict#" }, [app]);
 const suspendableApp = React.createElement(Suspense, { fallback: fallbackUi }, [strictApp]);
-const root = document.getElementById("reactroot");
 
-ReactDOM.render(suspendableApp, root);
+const container = document.getElementById("reactroot")!;
+const root = createRoot(container);
+
+root.render(suspendableApp);
